Create database with relations
