+++ database/requirements.txt
@@ -1,0 +1,2 @@
+sqlalchemy
+psycopg2-binary
